requirements.txt
# Core FastAPI dependencies
dynaconf==3.2.6
fastapi==0.115.5
starlette==0.41.3
uvicorn==0.32.1
pydantic==2.10.2

# Authentication and security
bcrypt==3.2.2
passlib==1.7.4
python-jose==3.3.0
cryptography==45.0.4
email-validator==2.0.0

# Database
motor==3.6.0
pymongo==4.9
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.29.0
psycopg2-binary==2.9.9

# LLM providers
# OpenAI
langchain==0.3.21
langchain_openai==0.2.10
openai==1.55.3

# Google Gemini
# langchain-google-genai>=0.0.6
langchain-core==0.1.18
google-ai-generativelanguage>=0.4.0
google-generativeai>=0.3.1

# Template and file handling
jinja2==3.1.5
aiofiles==23.2.1
python-multipart==0.0.18
python-dotenv==1.0.1

# Repository documentation
gitpython>=3.1.40
tqdm>=4.66.1

# Data processing
pandas==2.2.3
javalang==0.13.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
httpx>=0.24.0        # For async HTTP client in tests
faker>=18.0.0        # For generating test data
requests>=2.31.0
jsonschema==4.25.1
pyyaml==6.0.1
